fixed deployment issue and adjusted UI
