+++ app/(root)/cam/page.tsx
@@ -11,6 +11,7 @@ import Modal from "@/components/Modal";
 import { addToInventory } from "@/components/ImageUploader";
 import { CircularProgress } from "@mui/material";
 import "../../globals.css"
+import { parsePicDescription, renderParsedDescription } from "@/lib/helper";
 
 const Wrapper = styled.div`
   position: fixed;
@@ -150,25 +151,8 @@ const FullScreenImagePreview = styled.div<{ image: string | null }>`
   background-position: center;
 `;
 
-export const parsePicDescription = (description: string) => {
-  try {
-    const jsonStr = (description as any).object
-      .replace(/```json/g, "")
-      .replace(/```/g, "")
-      .replace(/\n/g, "")
-      .trim();
 
-    return JSON.parse(jsonStr);
-  } catch (error) {
-    console.error("Error parsing JSON:", error);
-    return null;
-  }
-};
 
-export const renderParsedDescription = (parsedDescription: any) => {
-  if (parsedDescription?.error) return parsedDescription.error;
-  return parsedDescription?.sku;
-};
 
 const Cam = () => {
   const [numberOfCameras, setNumberOfCameras] = useState(0);
